moving ukulele chord elements into Chord
