+++ src/components/Chord/Fingers.jsx
