+++ src/components/Chord/Frets.jsx
