+++ src/components/Chord/Nut.jsx
