+++ src/components/Chord/Strings.jsx
